feat: Add support for using mutliple storage items
Since we are using Chrome sync storage, this has
restrictions on the amout of data per item (8kb)
and the number of items (512) that one extension
can store in sync storage. The code in this commit
manages these restictions while still allowing for
many loans to be ignored (100, 000 at the moment).