fix: Make ignore feature survive filter change
This commit fixes the problem of the ignore buttons
disappearing and hidden loans appear whenever the
loan filter settings were changed.

The solution involves setting up a MutationObserver
to watch for changes to the page, specifically when
the table of loans updates and fades back in. Then
we re-run the functions to render the ignore buttons,
add the listeners to the buttons and hide any loans.

Thanks go to @LoneRifle for suggesting the idea of
using the MutationObserver. That was a great help.

Resolves: #1